Revert "Tutorial"
