+++ client/configurations.json
@@ -1,7 +1,7 @@
 {
   "hexagon": {
-    "width": 900,
-    "height": 900,
+    "width": 1200,
+    "height": 1000,
     "layout": { "width": 6, "height": 6, "flat": false, "spacing": 1.02 },
     "origin": { "x": 0, "y": 0 },
     "map": "hexagon",
